added authors table to createTable

+++ thebookattic-backend/src/setup/createPostgreSQLTables.sql
@@ -19,7 +19,7 @@ create table authors
 	userid int not null
 	firstname varchar(25),
 	lastname varchar(25),
-	avgrating int not null,
+	avgrating float not null,
 	bio varchar(500) not null,
 	picture varchar(100) not null
 )
